+++ docnexus/__init__.py
@@ -2,7 +2,7 @@
 DocNexus - The Ultimate All-in-One Document Engine
 """
 
-__version__ = '1.2.0'
+from .version_info import __version__
 __author__ = 'Doc Nexus Maintainers'
 __email__ = 'maintainers@docnexus.org'
 __description__ = 'DocNexus - The Ultimate All-in-One Document Engine'